+++ adventofcode/2022/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.5-beta.2](https://github.com/clabroche/monorepo/compare/v1.1.5-beta.1...v1.1.5-beta.2) (2022-12-14)
+
+### Bug Fixes
+
+* **adventofcode:** add day 14 ([b135242](https://github.com/clabroche/monorepo/commit/b135242318b9c493b77ea44da4ed9e481e63fda8))
+* **adventofcode:** add readme to day 10 ([1ddbf40](https://github.com/clabroche/monorepo/commit/1ddbf40c3dade79c8d29660021a7bdafee66050e))
+* **adventofcode:** add readme to day 11 ([8161b3d](https://github.com/clabroche/monorepo/commit/8161b3dd3c0a64f72a3914ba1e84f723707c3adb))
+* **adventofcode:** add readme to day 13 ([2dee171](https://github.com/clabroche/monorepo/commit/2dee1711841246946b1a36c0097c9d9621a798e0))
+
+## [1.1.5-beta.0](https://github.com/clabroche/monorepo/compare/v1.1.2-beta.11...v1.1.5-beta.0) (2022-12-13)
+
+### Bug Fixes
+
+* **adventofcode:** add day 10 ([4c18c41](https://github.com/clabroche/monorepo/commit/4c18c41b3b47627831a13b034dbee263d5575acc))
+* **adventofcode:** add day 11 ([15c4b98](https://github.com/clabroche/monorepo/commit/15c4b9898fbcde9bd684a18ee4ec04a64bddd352))
+* **adventofcode:** add day 12 ([d608419](https://github.com/clabroche/monorepo/commit/d608419571a16e0bbda226adadf803e79a099098))
+* **adventofcode:** add day 13 ([111ba48](https://github.com/clabroche/monorepo/commit/111ba48b7cee8f7e3bdafc610780a4b814dbfc81))
+* **adventofcode:** add day 5 ([8ccc1f4](https://github.com/clabroche/monorepo/commit/8ccc1f41e0101e0155316c5b14d1d562b89b9b6c))
+* **adventofcode:** add day 6 ([a85a7db](https://github.com/clabroche/monorepo/commit/a85a7dbfc591deb3193299bb911379717c9c76cd))
+* **adventofcode:** add day 7 ([af35806](https://github.com/clabroche/monorepo/commit/af358063f74532bb626088adce9035e34bd5ad69))
+* **adventofcode:** add day 8 ([67d9850](https://github.com/clabroche/monorepo/commit/67d9850422ec908171d2c00fb3d986ef4c37033b))
+* **adventofcode:** add day 9 ([9564177](https://github.com/clabroche/monorepo/commit/9564177d05bbecfda7c03086164b2ca48fa0c663))
+* **adventofcode:** remove unecessary code from 2022-12-01 ([1f7a42a](https://github.com/clabroche/monorepo/commit/1f7a42a5c38ff788d517bd84190b9b93a278903b))
+
+## [1.1.4](https://github.com/clabroche/monorepo/compare/v1.1.4-beta.5...v1.1.4) (2022-12-12)
+
+**Note:** Version bump only for package @clabroche/adventofcode-2022
+
+## [1.1.4-beta.5](https://github.com/clabroche/monorepo/compare/v1.1.4-beta.4...v1.1.4-beta.5) (2022-12-12)
+
+### Bug Fixes
+
+* **adventofcode:** add day 10 ([e896e55](https://github.com/clabroche/monorepo/commit/e896e55554aa157aea3be7cfb2fe3566c6321fff))
+* **adventofcode:** add day 11 ([dec0668](https://github.com/clabroche/monorepo/commit/dec0668575301b90d2ea85b4c8104e66d56ec420))
+* **adventofcode:** add day 9 ([72e1f2e](https://github.com/clabroche/monorepo/commit/72e1f2e7834e72a4a7f27c12e82ec07683690e98))
+
+## [1.1.4-beta.4](https://github.com/clabroche/monorepo/compare/v1.1.4-beta.3...v1.1.4-beta.4) (2022-12-08)
+
+### Bug Fixes
+
+* **adventofcode:** add day 8 ([9f221d9](https://github.com/clabroche/monorepo/commit/9f221d960fe14c2cf23563e016a2edd87decc6e5))
+
 ## [1.1.4-beta.3](https://github.com/clabroche/monorepo/compare/v1.1.4-beta.2...v1.1.4-beta.3) (2022-12-07)
 
 ### Bug Fixes